Renamed older components
